% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfb.R
\name{sfb}
\alias{sfb}
\title{#' Sectoral Balances
#'
#' This function receives AMECO sectoral balances data and outputs data and plots
#'}
\usage{
sfb(list_of_countries, t_start = 1950, t_end = 9999, getData = F, getPlot = T)
}
\arguments{
\item{list_of_countries}{List of countries}

\item{t_start}{start year}

\item{t_end}{end year}

\item{getData}{output data table}

\item{getPlot}{output sectoral balance plot}
}
\value{
data and/or plots of sectoral balances
}
\description{
#' Sectoral Balances
#'
#' This function receives AMECO sectoral balances data and outputs data and plots
#'
}
